requirements.txt
amqp==1.4.9
anyjson==0.3.3
appdirs==1.4.0
apt-xapian-index==0.47
apturl==0.5.2
beautifulsoup4==4.4.1
billiard==3.3.0.23
bleach==1.5.0
blinker==1.3
Brlapi==0.6.4
cart==1.1.3
celery==3.1.25
chardet==2.3.0
checkbox-support==0.22
colorama==0.3.9
command-not-found==0.3
cryptography==1.2.3
cycler==0.10.0
defer==1.0.6
Django==1.9
django-background-tasks==1.1.13
django-bootstrap-datepicker==1.2.3
django-braces==1.12.0
django-cart==1.0.4
django-celery==3.2.2
django-compat==1.0.15
django-cors-headers==2.1.0
django-crispy-forms==1.7.0
django-datetime-widget==0.9.3
django-floppyforms==1.7.0
django-froala-editor==2.7.5
django-image-compressor==0.0.4
django-ipware==2.0.1
django-jet==1.0.7
django-markdown-deux==1.0.5
django-model-utils==3.1.1
-e git+https://timmyomahony@github.com/timmyomahony/django-pagedown.git@3ca542297733065444c80bdf4974805d8a721837#egg=django_pagedown
django-password-reset==1.0
django-social-widgets==0.4.0
django-star-ratings==0.6.0
django-widget-tweaks==1.4.1
django-wpadmin==1.7.4
djangorestframework==3.7.7
easy-thumbnails==2.5
enum34==1.1.6
feedparser==5.1.3
guacamole==0.9.2
html5lib==0.9999999
httplib2==0.9.1
idna==2.0
Jinja2==2.8
kombu==3.0.37
language-selector==0.1
louis==2.6.4
lxml==3.5.0
Mako==1.0.3
Markdown==2.6.11
markdown2==2.3.5
MarkupSafe==0.23
matplotlib==2.1.2
mysqlclient==1.3.12
numpy==1.14.0
oauthlib==1.0.3
olefile==0.44
onboard==1.2.0
packaging==16.8
padme==1.1.1
pandas==0.22.0
pexpect==4.0.1
Pillow==5.0.0
plainbox==0.25
protobuf==3.5.1
ptyprocess==0.5
pyasn1==0.1.9
pycryptodome==3.4.8
pycups==1.9.73
pycurl==7.43.0
pygobject==3.20.0
PyJWT==1.3.0
PyMySQL==0.8.0
pyparsing==2.1.10
python-apt==1.1.0b1
python-dateutil==2.6.1
python-debian==0.1.27
python-systemd==231
pytz==2017.3
pyxdg==0.25
regex==2018.1.10
reportlab==3.3.0
requests==2.9.1
scipy==1.0.0
sessioninstaller==0.0.0
six==1.10.0
ssh-import-id==5.5
swapper==1.1.0
system-service==0.3
tensorflow==1.4.1
tensorflow-tensorboard==0.4.0rc3
tqdm==4.19.5
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-gdrive==0.7
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urllib3==1.13.1
usb-creator==0.3.0
vine==1.1.4
virtualenv==15.0.1
Werkzeug==0.14.1
xdiagnose==3.8.4.1
xkit==0.0.0
XlsxWriter==0.7.3
